requirements.txt
asgiref==3.7.2
attrs==22.2.0
certifi==2024.2.2
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==2023.10
crispy-tailwind==1.0.3
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-email==0.1.10
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
exceptiongroup==1.1.1
filelock==3.13.1
Flask==2.2.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.6
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.26.3
packaging==23.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest==7.2.2
pytweening==1.0.7
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.0
Werkzeug==2.2.2